now correct
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/trunk/kdchart@673 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ examples/Bars/mainwindow.cpp
@@ -113,7 +113,7 @@ void MainWindow::on_paintValuesCB_toggled( bool checked )
 void MainWindow::on_paintThreeDBarsCB_toggled( bool checked )
 {
     ThreeDBarAttributes td;
-    int defaultDepth = static_cast<int>(td.depth());
+    double defaultDepth = td.depth();
     if ( checked ) {
         td.setEnabled( true );
         if ( threeDDepthCB->isChecked() )
